+++ CHANGELOG.md
@@ -14,7 +14,14 @@
 * YamlProvider now supports a `shared_filename` that can be used to add a set of
   common records across all zones using the provider. It can be used stand-alone
   or in combination with zone files and/or split configs to aid in DRYing up DNS
-  configs.
+* YamlProvider now supports an `!include` directive which enables shared
+  snippets of config to be reused across many records, e.g. common dynamic rules
+  across a set of services with service-specific pool values or a unified SFP
+  value included in TXT records at the root of all zones.
+* SpfRecord is formally deprecated with an warning and will become a
+  ValidationError in 2.x
+* SpfDnsLookupProcessor is formally deprcated in favor of the version relocated
+  into https://github.com/octodns/octodns-spf and will be removed in 2.x
 * MetaProcessor added to enable some useful/cool options for debugging/tracking
   DNS changes. Specifically timestamps/uuid so you can track whether changes
   that have been pushed to providers have propogated/transferred correctly.